requirements.txt
fastapi
uvicorn
python-multipart
twilio
groq
supabase
python-dotenv
httpx
